+++ VisualChatGPTStudioShared/Commands/AddComments.cs
@@ -1,6 +1,7 @@
 ﻿using Community.VisualStudio.Toolkit;
 using JeffPires.VisualChatGPTStudio.Commands;
 using JeffPires.VisualChatGPTStudio.Options.Commands;
+using JeffPires.VisualChatGPTStudio.Utils;
 
 namespace JeffPires.VisualChatGPTStudio
 {